+++ llvm/lib/Target/AMDGPU/SIRegisterInfo.cpp
@@ -4006,13 +4006,3 @@ SIRegisterInfo::getSubRegAlignmentNumBits(const TargetRegisterClass *RC,
   }
   return 0;
 }
-
-SmallVector<StringLiteral>
-SIRegisterInfo::getVRegFlagsOfReg(Register Reg,
-                                  const MachineFunction &MF) const {
-  SmallVector<StringLiteral> RegFlags;
-  const SIMachineFunctionInfo *FuncInfo = MF.getInfo<SIMachineFunctionInfo>();
-  if (FuncInfo->checkFlag(Reg, AMDGPU::VirtRegFlag::WWM_REG))
-    RegFlags.push_back("WWM_REG");
-  return RegFlags;
-}